Simplify logic with NSPersistentContainer (requires 10.12)
Remove the _store and _coordinator ivars in favor of an
NSPersistentContainer object that manages both. There are three
notable changes to the API here:

* managedObjectContextClass has been removed, and managedObjectContext
  is now read-only. The setter method was used for custom subclasses,
  but the Apple docs state in no uncertain terms "You are strongly
  discouraged from subclassing NSManagedObjectContext." So the loss is
  probably not significant.

* The overridable method

  - (BOOL)configurePersistentStoreCoordinatorForURL:(NSURL *)storeURL
                                             ofType:(NSString *)fileType
                                 modelConfiguration:(NSString *)configuration
                                       storeOptions:(NSDictionary *)storeOptions
                                              error:(NSError **)error;

  has been supplemented with an alternative method

  - (BOOL)configurePersistentStoreCoordinatorWithDescription:(NSPersistentStoreDescription *)description
                                                       error:(NSError **)outError;

  Subclasses can mutate the NSPersistentStoreDescription before calling
  super. This simplifies the process e.g. of setting SQLite pragmas.
  Internally, the first method calls the second so existing code
  overrides that call super will behave as expected.

* BSManagedDocument now requires 10.12 or later, as
  NSPersistentContainer was introduced in 10.12.

The motivation behind this change is three-fold. First, it allows us to
get rid of delicate code paths with several-hundred-word code comments
and defer to Apple's implementation for managing the stores and
contexts. Second, it opens the door to supporting Core Spotlight
directly in BSManagedDocument (see NSCoreDataCoreSpotlightDelegate
introduced in 10.13, which works with NSPersistentStoreDescriptions).
Third, configurePersistentStoreCoordinatorWithDescription:error: allows
for somewhat simpler subclass code.

Some of the old logic could be restored to continue to support 10.8 -
10.11. However, my requirements no longer mandate support going that
far back, and so this change also includes a few simplifications and
removal of pre-10.12 code paths and comments.