Fix "File modified by another application" error
Prior to NSPersistentContainer we had a main context (on the main thread)
and a parent context (on a background thread), both of which needed to
be saved.

With NSPersistentContainer, the main context no longer has a parent
context (parentContext returns nil). Previously I added a private
context to do the saving, but this doesn't make any sense because the
private context won't have the main context's changes. Using the main
context for the second save results in a file-modification error.

Here I do the implement the naive solution and simply remove the
two-stage save. Saving only the main (view) context appears to propagate
the save up to the store properly. It's possible that our friend the
container is managing a private context (i.e. the former parent context)
but I'm not sure. If there is no private context traipsing about behind
the NSPersistentContainer scenes, then it's possible that asynchronous
saving will not work under this architecture. More testing is needed to
find out.