Fix unsaved document behavior
NSPersistentContainer doesn't like it if you try to access a context
before a persistent store has been configured. In the old way of doing
things, an unsaved document would have an unaffiliated context, and
saving the document would associate the existing context with the chosen
persistent store. This doesn't work with NSPersistentContainer, which
emits a warning when accessing the viewContext when no persistent store
is present, and those contexts will return errors whenever a save: is
attempted.

As a workaround, when the document's managedObjectContext is requested,
test to see if a persistent store has been configured. If not, force a
synchronous autosave, which will add an autosaved persistent store.
Then return a store-backed context that is ready for action.