File duplication fixes
Duplication is implemented inside NSDocument as an AutosaveElsewhere, a
fact that I did not appreciate earlier. This explains why the original
code tested inAbsoluteURL for equality with the autosavedContentsFileURL
before using the custom writing logic - we want the "regular channels"
during file duplication.

I've supplemented the original equality test with a check on
mostRecentlySavedFileURL, which will only be null for unsaved,
un-autosaved documents. This should pick up on the early-forced-autosave
that is used during context creation, but pass on file duplication
either of an unsaved document (which will have an autosave URL thanks
to its own forced autosave) or of a saved document (which will have
a fileURL).