Implement Share for unsaved documents
Duplicate and Share are very similar under the hood - they end up
calling writeSafelyToURL: with a save operation type of
NSAutosaveElsewhereOperation and a destination URL that has nothing
to do with the regular `autosavedContentsFileURL`.

Previously, when autosavedContentsFileURL wasn't equal to the destination,
we'd "fall back to the regular channels" and call super's
writeSafelyToURL:. This relied on having a writingBlock that knew it
should do a dumb copy instead of trying to commandeer the persistent
store URL. This worked fine for Duplicate, which knew to set a
writingBlock, but did not work for Share, which had no idea what was
expected of it.

I've refactored things a set the file-copy writingBlock inside
of writeSafelyToURL: whenever the autosave elsewhere destination
is not the regular autosavedContentsFileURL. This happens for Duplicate
and for Share, and who knows, maybe Apple will think of other ways
to abuse Autosave in the future. I have simplified the logic of
duplicateAndReturnError:, and have overridden
shareDocumentWithSharingService:completionHandler: to mirror its
three basic steps:

1. Ensure a backing store exists
2. Perform a synchronous autosave
3. Call super

Seems to work with basic testing but all file functions will need to
be put through their paces since writeSafelyToURL: has been modified.