Fix a deadlock and explain it with an extensive comment
Disable asynchronous writing until the persistent container is fully
configured. This is necessary because to retrieve a managedObjectContext
for the first time, we need to block while waiting for an autosave to
complete, but the autosave needs access to the main thread to do a save:
on the created viewContext, resulting in a deadlock.

It seems to me that as long as we're providing the container's
viewContext to subclasses, save: will have to occur on the main thread,
which negates most of the benefit of async writing. Maybe someone can
figure out how to shuffle the changes between contexts before writing
to disk (e.g. with mergeChangesFromContextDidSaveNotification:), but in
the meantime this change fixes the deadlock.