Fix a deadlock with file duplication
The async file saving resulted in another deadlock during the forced
autosave initiated by a file duplication. Resolve it with judicious
use of continueAsynchronousWorkOnMainThreadUsingBlock:.

It's possible that with this change we can go back to always returning
YES from canAsynchronouslyWriteToURL:, since the file-creation deadlocks
occured for the same reason (synchronous file activity waiting for the
autosave to complete).