Remove the long "sadness and wretchedness" code path
When reverting a document with unautosaved changes, NSDocument
will attempt to preserve a version prior to calling
revertToContentsOfURL:ofType:error:. It looks like it is sufficient
to set the autosavedContentsFileURL to the temporary location in
order to get the version saved to the correct place. The regular
machinery will take care of creating a package directory and
shooting out a copy to the correct store URL, so we don't need
the long code path that performs these steps manually.

Maybe I'm missing something, but this seems to work just fine and
lets us rip out a lot of code.